Separate Gravity into two modules, adding GravityWithCenter
