+++ modular-particle-system/CHANGELOG.md
@@ -12,6 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added optional `repeat` property to all generators `bursts` properties.
     -   Can be used to make bursts repeat indefinitely.
 
+-   Added `GravityWithCenter` modifier.
+
+### Changed
+
+-   Removed `center` and related properties from `Gravity` modifier.
+    -   These were separated to their own module `GravityWithCenter`.
+
 ## [1.0.0] - 2022-04-10
 
 ### Added
